pyproject.toml
[project]
name = "strauss"
version = "0.1.0"
description = "Candidate-advocate agent for Bryan Zane Smith — multi-provider (Anthropic, OpenAI, Moonshot Kimi K2)."
readme = "README.md"
requires-python = ">=3.11"
authors = [{ name = "Bryan Zane Smith", email = "bzane09@gmail.com" }]
license = { text = "MIT" }

dependencies = [
    "anthropic>=0.40.0",
    "openai>=1.50.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["backend"]
